+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased]
 
+### Fixed
+
+- When cache_all_requests was on, some requests were not being cached
+
 ## [0.5.0] - 2019-05-09
 
 ### Added